Updated repository cpan metadata & removed bugtracker
Removed bugtracker since it was in the wrong place (should have been with repository in resources), it appears that there is a move away from sourceforge in general for the project and RT is doing the job handling tickets anyways.  There are 9 open bug still in the sourceforge bug tracker so there should likely be a process of of seeing if they are still issues and having them moved to RT.  For instance you could ask the submitters to do the checking and resubmitting.