Skip XML::Parser::Lite tests if it is missing
We simply try to load it in an eval, and if that fails, we skip the second half
of the tests.

This should make it possible to install SOAP::Lite on a fresh system again,
since right now there is a circular dep between SOAP::Lite and
XML::Parser::Lite, resulting in uninstallable modules unless you force one or
the other.