Update Lite.pm
It was necessary in one project, that i can add attributes to the Header and Body tags in an RPC call. I couldn't find any way, to do that, except changing the serializer of SOAP::Lite.
Overriding the method SOAP::Serializer::envelope seemed to be an overkill and would have produced incompatibilities with future versions of SOAP::Lite, i decided to do the change in SOAP::Lite itself, hoping, it will be accepted.

The names headerattr and bodyattr appeared ok for me, but are just a proposal.